Decouple OpenAI client and Vortex
